fixed zoom mode error - move when safe.

+++ lib/goom/src/goom_graphic.h
@@ -7,9 +7,10 @@
 
 struct Color
 {
-  uint8_t r;
-  uint8_t g;
-  uint8_t b;
+  // TODO: Why uint16 here and not uint8?
+  uint16_t r;
+  uint16_t g;
+  uint16_t b;
 };
 
 extern const Color BLACK;
